めぶきチャンネル対応
