+++ src/index.html
@@ -6,7 +6,31 @@
 <title>DiscoTegaki - 手書き</title>
 <link rel="icon" href="/discord-tegaki/asset/favicon.ico">
 <link href="./style.css" rel="stylesheet">
-<script src="./main.js" defer></script>
+<script src="./app.js" defer></script>
+<script>
+document.addEventListener("DOMContentLoaded", async () => {
+  const app = await DiscoTegaki.launch("https://toshiakidisco.github.io/discord-tegaki/");
+  app.disableClose();
+  app.open(0, 0);
+});
+</script>
+
+<style>
+
+html, body {
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  min-height: 100vh;
+  position: relative;
+}
+
+</style>
+
 <body class="github-pages">
 </body>
 </html>
