+++ src/module.ts
@@ -1,0 +1,9 @@
+import { DiscordTegaki } from "./app";
+
+declare global {
+  interface Window {
+    DiscoTegaki: typeof DiscordTegaki;
+  }
+}
+
+window["DiscoTegaki"] = DiscordTegaki;
